Move enemy and projectile component to turret from turret_attack
